ARZNI subproject updated from commit `b93c897` to `5b1d83e-dirty`
This commit message describes the code diff between two commits of a subproject named ARZNI in a directory with a Cyrillic name (in this case, "Общая информация/Арзни"). The subproject was updated from commit `b93c897` to commit `5b1d83e-dirty`. The "-dirty" suffix indicates that there are uncommitted changes in the working directory of the subproject.